Windows MSVC: 64-bit <time.h> functions, and tests that hold on LLP64
`time`, `difftime`, `mktime`, `localtime`, `gmtime` and `ctime` linked on
MSVC — as weak aliases for the *32-bit* `_time32` family in msvcrt.lib,
while the bundled <time.h> makes `time_t` eight bytes as the UCRT does.
So `mktime` of an impossible date read back as 4294967295 and `gmtime`
past 2038 was NULL, silently; `timespec_get` did not link at all. On an
MSVC target a declared-only function now links by the name the UCRT
exports (`_time64` … `_timespec64_get`, and `_hypotf` for `hypotf`); an
asm label still wins, a definition is untouched, mingw-w64 is left alone.
`fabsf`, `frexpf`, `ldexpf` and the `wmem*` functions have no exported
counterpart and are documented with workarounds.

Everything else found by running on Windows for the first time was the
tests assuming LP64: `long` for a pointer or a `sizeof`, `int` for
`wchar_t`, `1 << 40` in a `c_long`. They use `uintptr_t`, `size_t`,
`wchar_t` and `long long` now, or follow `c_long`'s width; glibc-only and
POSIX-only tests are gated to where they apply. No translation bug was
found on LLP64: 36 test files and both examples build, link and run on
x86_64-pc-windows-msvc.

- scripts/portability-tests.txt is the list the CI portability job and
  scripts/test-windows-from-wsl.sh both read; the latter builds with
  lld-link against the Windows host's MSVC and SDK libraries and runs the
  result through WSL interop
- tests/portability.rs checks the calendar functions field by field,
  a pre-epoch `mktime` included, on every platform

+++ tests/aggregates.rs
@@ -884,7 +884,16 @@ fn integer_constants_cast_to_a_function_pointer_type() {
     // A non-zero constant is an implementation-defined conversion, and goes
     // through `usize`, since the integer type C names need not be
     // pointer-sized. (c-testsuite 00159.)
+    //
+    // The round trip goes through `uintptr_t` rather than `unsigned long`,
+    // which is the type C has for "an integer a pointer survives": `long` is
+    // four bytes on Windows and would truncate the address, and a truncated
+    // function pointer called is an access violation rather than a test
+    // failure. `a_narrow_constant` below is the case where the integer is
+    // deliberately *not* wide enough.
     c99! {
+        #include <stdint.h>
+
         typedef void (*Action)(void);
         typedef int (*Unary)(int);
 
@@ -906,8 +915,8 @@ fn integer_constants_cast_to_a_function_pointer_type() {
             return (a == 0) + (b == 0) + (c == 0);
         }
 
-        Unary from_integer(unsigned long bits) { return (Unary) bits; }
-        unsigned long to_integer(Unary f) { return (unsigned long) f; }
+        Unary from_integer(uintptr_t bits) { return (Unary) bits; }
+        uintptr_t to_integer(Unary f) { return (uintptr_t) f; }
         int roundtrip(int n) { return from_integer(to_integer(negate))(n); }
 
         int a_narrow_constant(void) {
